add canStart (game can't start if there's only one player)

+++ src/main/java/fr/leroideskiwis/uno/managers/GameManager.java
@@ -13,10 +13,12 @@ public class GameManager {
     private final List<Game> games = new ArrayList<>();
 
     public void startGame(Group group, GroupManager groupManager){
-        Game game = group.toGame();
-        games.add(game);
-        game.startGame();
-        groupManager.deleteGroup(group);
+        if(group.canStart()) {
+            Game game = group.toGame();
+            games.add(game);
+            game.startGame();
+            groupManager.deleteGroup(group);
+        }
 
     }
